feat: 서버 사이드에서 data fetch
