+++ frontend/techpick/src/providers/QueryProvider.tsx
@@ -1,34 +1,9 @@
 'use client';
 
-import {
-  QueryClient,
-  QueryClientProvider,
-  isServer,
-} from '@tanstack/react-query';
+import { getQueryClient } from '@/libs/@react-query/getQueryClient';
+import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import type { PropsWithChildren } from 'react';
-
-const makeQueryClient = () => {
-  return new QueryClient({
-    defaultOptions: {
-      queries: {
-        staleTime: 60 * 1000,
-      },
-    },
-  });
-};
-
-let browserQueryClient: QueryClient | undefined = undefined;
-
-const getQueryClient = () => {
-  if (isServer) {
-    return makeQueryClient();
-  }
-  if (!browserQueryClient) {
-    browserQueryClient = makeQueryClient();
-  }
-  return browserQueryClient;
-};
 
 export const QueryProvider = ({ children }: PropsWithChildren) => {
   const queryClient = getQueryClient();
